+++ include/cse/async_slave.hpp
@@ -5,12 +5,15 @@
 #ifndef CSE_ASYNC_SLAVE_HPP
 #define CSE_ASYNC_SLAVE_HPP
 
+#include <memory>
 #include <string>
+#include <string_view>
 
 #include <boost/fiber/future/future.hpp>
 #include <gsl/span>
 
 #include <cse/model.hpp>
+#include <cse/slave.hpp>
 
 
 namespace cse
@@ -164,6 +167,16 @@ public:
         gsl::span<const variable_index> stringVariables,
         gsl::span<const std::string> stringValues) = 0;
 };
+
+
+/**
+ *  Wraps a synchronous slave in an asynchronous interface.
+ *
+ *  The resulting `async_slave` is not actually asynchronous, as all function
+ *  calls will be executed in fibers in the current thread (hence the "pseudo"
+ *  modifier).
+ */
+std::unique_ptr<async_slave> make_pseudo_async(std::unique_ptr<slave> s);
 
 
 } // namespace cse
